Update README.md
This is my first ever pull request, apologies in advance if I've messed up somewhere and somehow, I think I've got it right though.

I've added a prompt that I've been using successfully in Thailand for week so.  It's a little lengthy, but served very well so far.  A legacy translations apps have always been poor at Latin2SEAsiaScript translations which is one reason for the size, the other is there's several ways to 'act' here depending on familiarity, but as this is likely to be used with people you don't know so well I've tried to ensure a respectful output.  A few people here have been using it and so far the feedback has been overwelmingly positive.